Remove Ruby 1.9 support for method instrumentation
In PR #691 and #692 I added support for method instrumentation with
keyboard arguments for Ruby 3.0. This included support also for Ruby
1.9, with a less than ideal solution of requiring different
implementation files based on the language version for syntax
compatibility reasons.

Since the develop branch for the Ruby gem 3.0 release drops Ruby 1.9
support, undo the exceptions made for Ruby 1.9 support and only keep the
"modern Ruby" implementation moving forward.
